Add timestamp comparison

+++ main.py
@@ -22,6 +22,7 @@ router = APIRouter()
 class QueryInfoBase(BaseModel):
     id: str
     id_column: str
+    timestamp: str
     timestamp_column: str
     tables: List[str]
     statement: str
@@ -38,8 +39,11 @@ async def update_records(query_info: QueryInfoBase, db: DbPoolConnAndCursor = De
     await cursor.execute(f"select {query_info.timestamp_column} from {query_info.tables[0]} WHERE {query_info.id_column} = '{query_info.id}'")      
     existing_record = await cursor.fetchone()
     if existing_record:
-        for table in query_info.tables:
-            await cursor.execute(f"DELETE FROM {table} where {query_info.id_column} = '{query_info.id}'")
+        if existing_record[query_info.timestamp_column] <= query_info.timestamp:
+            for table in query_info.tables:
+                await cursor.execute(f"DELETE FROM {table} where {query_info.id_column} = '{query_info.id}'")
+        else:
+            return {}
     await cursor.execute(query_info.statement)
     return {}
